Refactor / Rename enum package to parameters
