+++ .idea/workspace.xml
@@ -8,7 +8,16 @@
       <change afterPath="$PROJECT_DIR$/MIT.md" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/README.md" beforeDir="false" afterPath="$PROJECT_DIR$/README.md" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/org/example/domain/Bedclothing.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/org/example/domain/Bedclothing.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/main/java/org/example/domain/Dish.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/org/example/domain/Dish.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/org/example/domain/Inventory.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/org/example/domain/Inventory.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/org/example/domain/enums/BeddingSize.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/org/example/domain/parameters/BeddingSize.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/org/example/domain/enums/DishSize.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/org/example/domain/parameters/DishSize.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/org/example/domain/enums/FillingMaterial.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/org/example/domain/parameters/FillingMaterial.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/org/example/domain/enums/FoodSafety.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/org/example/domain/parameters/FoodSafety.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/org/example/domain/enums/HeatResistance.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/org/example/domain/parameters/HeatResistance.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/org/example/domain/enums/Material.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/org/example/domain/parameters/Material.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/org/example/domain/enums/Style.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/org/example/domain/parameters/Style.java" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -66,7 +75,7 @@
       <updated>1684823110604</updated>
       <workItem from="1684823111625" duration="1246000" />
       <workItem from="1684840088470" duration="612000" />
-      <workItem from="1685294356299" duration="8418000" />
+      <workItem from="1685294356299" duration="8468000" />
     </task>
     <servers />
   </component>
